make `code2md` bash 3 compatible
fucking fsf, ruining macos...

+++ kernel/raspi/uart0.cc
@@ -17,10 +17,10 @@
 void uart_init() {
         int r;
 
-        /* initialize UART */
+        // initialize UART
         UART0_CR = 0; // turn off UART0
 
-        /* set up clock for consistent divisor values */
+        // set up clock for consistent divisor values
         mbox[0] = 9 * 4;
         mbox[1] = MBOX_REQUEST;
         mbox[2] = MBOX_TAG_SETCLKRATE; // set clock rate
@@ -32,7 +32,7 @@ void uart_init() {
         mbox[8] = MBOX_TAG_LAST;
         mbox_call(MBOX_CH_PROP);
 
-        /* map UART0 to GPIO pins */
+        // map UART0 to GPIO pins
         r = GPFSEL1;
         r &= ~((7 << 12) | (7 << 15)); // gpio14, gpio15
         r |= (4 << 12) | (4 << 15);    // alt0
@@ -51,7 +51,7 @@ void uart_init() {
         UART0_CR   = 0x301;     // enable Tx, Rx, FIFO
 }
 
-void uart_putc(int c) {
+void uart_putc(int_fast8_t c) {
         if(!c) return;
         spinwhile(UART0_FR & 0x20);
         UART0_DR = c;
