+++ kernel/raspi/uart0.hh
@@ -1,10 +1,11 @@
 #pragma once
+#include <stdint.h>
 
 // Initializes UART0.
 void uart_init();
 
 // putc over UART0
-void uart_putc(int c);
+void uart_putc(int_fast8_t c);
 
 // getc over UART0
 int uart_getc();
